\name{readdata}
\alias{readdata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{import data}
\description{
After we collect data with Google Sheet,we need to summarize and
analyse it with r,so we can use this function to import data with
the link of our Google Sheet provide when we publish it on web.
}
\usage{
readdata(link)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{link}{
the link of the Google Sheet provide when we publish the sheet in the kind of .CSV on web.
}
}
\value{
data you collected in the google sheet.
}

\author{"Sheng Wang"}
