\name{scatterplot}
\alias{scatterplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
    show the attributes of our data in scatterplot.
}
\description{
    when we get data initially, it is difficult for us to find the attributes of data           diretly, so we can use the scatterplot to express the paired data in scatterplot, so        that we can get the most common data and the biggest frequency of range.
}
\usage{scatterplot(data,x_data,y_data,x_lab,y_lab,title)
}
\arguments{
  \item{data}{data we have already collected.
}
  \item{x_data}{data we set to be x-axis
}
  \item{y_data}{data we set to be y-axis
}
  \item{x_lab}{the label of x-axis
}
  \item{y_lab}{the label of y-axis
}
  \item{title}{the title of this scatterplot
}
}
\value{ the scatterplot using data data_x and data_y
}
\author{
Sheng Wang
}
